redistribute the functionalities in `support.py` and `kernel.py` modules to the `bank_erosion.utils.py` and the general `utils.py`
