+++ src/dfastbe/bank_erosion/utils.py
@@ -18,7 +18,7 @@ from dfastbe.bank_erosion.data_models import (
 )
 from dfastbe.io.logger import log_text
 from dfastbe.io.data_models import LineGeometry
-from dfastbe.support import enlarge, get_slices, get_slices_core, on_right_side
+from dfastbe.utils import on_right_side
 
 EPS = sys.float_info.epsilon
 water_density = 1000  # density of water [kg/m3]
@@ -1079,3 +1079,612 @@ def write_km_eroded_volumes(stations: np.ndarray, volume: np.ndarray, file_name:
         for i in range(len(stations)):
             str_value = "\t".join(["{:.2f}".format(x) for x in volume[i, :]])
             file.write("{:.2f}\t".format(stations[i]) + str_value + "\n")
+
+
+def get_slices(
+    index: int,
+    prev_b: float,
+    bpj: np.ndarray,
+    bpj1: np.ndarray,
+    mesh_data: MeshData,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """Calculate the intersection of a line segment with the edges of a mesh face.
+
+    This function determines where a line segment (defined by two points) intersects the edges of a mesh face.
+    It returns the relative distances along the segment and the edges where the intersections occur, as well as
+    flags indicating whether the intersections occur at nodes.
+
+    Args:
+        index (int):
+            Index of the current mesh face. If `index` is negative, the function assumes the segment intersects
+            the boundary edges of the mesh.
+        prev_b (float):
+            The relative distance along the previous segment where the last intersection occurred. Used to filter
+            intersections along the current segment.
+        bpj (np.ndarray):
+            A 1D array of shape (2,) containing the x, y coordinates of the current point of the line segment.
+        bpj1 (np.ndarray):
+            A 1D array of shape (2,) containing the x, y coordinates of the previous point of the line segment.
+        mesh_data (MeshData):
+            An instance of the `MeshData` class containing mesh-related data, such as edge coordinates, face-edge
+            connectivity, and edge-node connectivity.
+
+    Returns:
+        Tuple[np.ndarray, np.ndarray, np.ndarray]:
+            A tuple containing:
+            - `b` (np.ndarray): Relative distances along the segment `bpj1-bpj` where the intersections occur.
+            - `edges` (np.ndarray): Indices of the edges that are intersected by the segment.
+            - `nodes` (np.ndarray): Flags indicating whether the intersections occur at nodes. A value of `-1`
+              indicates no intersection at a node, while other values correspond to node indices.
+
+    Raises:
+        ValueError:
+            If the input data is invalid or inconsistent.
+
+    Notes:
+        - If `index` is negative, the function assumes the segment intersects the boundary edges of the mesh.
+        - The function uses the `get_slices_core` helper function to calculate the intersections.
+        - Intersections at nodes are flagged in the `nodes` array, with the corresponding node indices.
+
+    """
+    if index < 0:
+        edges = mesh_data.boundary_edge_nrs
+    else:
+        edges = mesh_data.face_edge_connectivity[index, : mesh_data.n_nodes[index]]
+    a, b, edges = get_slices_core(edges, mesh_data, bpj1, bpj, prev_b, True)
+    nodes = -np.ones(a.shape, dtype=np.int64)
+    nodes[a == 0] = mesh_data.edge_node[edges[a == 0], 0]
+    nodes[a == 1] = mesh_data.edge_node[edges[a == 1], 1]
+    return b, edges, nodes
+
+def get_slices_core(
+    edges: np.ndarray,
+    mesh_data: MeshData,
+    bpj1: np.ndarray,
+    bpj: np.ndarray,
+    bmin: float,
+    bmax1: bool = True,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """Calculate the intersection of a line segment with multiple mesh edges.
+
+    This function determines where a line segment intersects a set of mesh edges.
+    It calculates the relative distances along the segment and the edges where
+    the intersections occur, and returns the indices of the intersected edges.
+
+    Args:
+        edges (np.ndarray):
+            Array containing the indices of the edges to check for intersections.
+        mesh_data (MeshData):
+            An instance of the `MeshData` class containing mesh-related data,
+            such as edge coordinates and connectivity information.
+        bpj1 (np.ndarray):
+            A 1D array of shape (2,) containing the x, y coordinates of the
+            starting point of the line segment.
+        bpj (np.ndarray):
+            A 1D array of shape (2,) containing the x, y coordinates of the
+            ending point of the line segment.
+        bmin (float):
+            Minimum relative distance along the segment `bpj1-bpj` at which
+            intersections should be considered valid.
+        bmax1 (bool, optional):
+            If True, limits the relative distance along the segment `bpj1-bpj`
+            to a maximum of 1. Defaults to True.
+
+    Returns:
+        Tuple[np.ndarray, np.ndarray, np.ndarray]:
+            A tuple containing:
+            - `a` (np.ndarray): Relative distances along the edges where the
+              intersections occur.
+            - `b` (np.ndarray): Relative distances along the segment `bpj1-bpj`
+              where the intersections occur.
+            - `edges` (np.ndarray): Indices of the edges that are intersected
+              by the segment.
+
+    Raises:
+        ValueError:
+            If the input data is invalid or inconsistent.
+
+    Notes:
+        - The function uses the `get_slices_ab` helper function to calculate the
+          relative distances `a` and `b` for each edge.
+        - The `bmin` parameter is used to filter out intersections that occur
+          too close to the starting point of the segment.
+        - If `bmax1` is True, intersections beyond the endpoint of the segment
+          are ignored.
+    """
+    a, b, slices = _get_slices_ab(
+        mesh_data.x_edge_coords[edges, 0],
+        mesh_data.y_edge_coords[edges, 0],
+        mesh_data.x_edge_coords[edges, 1],
+        mesh_data.y_edge_coords[edges, 1],
+        bpj1[0],
+        bpj1[1],
+        bpj[0],
+        bpj[1],
+        bmin,
+        bmax1,
+    )
+    edges = edges[slices]
+    return a, b, edges
+
+
+def _get_slices_ab(
+    X0: np.ndarray,
+    Y0: np.ndarray,
+    X1: np.ndarray,
+    Y1: np.ndarray,
+    xi0: float,
+    yi0: float,
+    xi1: float,
+    yi1: float,
+    bmin: float,
+    bmax1: bool = True,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Get the relative locations a and b at which a segment intersects/slices a number of edges.
+
+    Arguments
+    ---------
+    X0 : np.ndarray
+        Array containing the x-coordinates of the start point of each edge.
+    X1 : np.ndarray
+        Array containing the x-coordinates of the end point of each edge.
+    Y0 : np.ndarray
+        Array containing the y-coordinates of the start point of each edge.
+    Y1 : np.ndarray
+        Array containing the y-coordinates of the end point of each edge.
+    xi0 : float
+        x-coordinate of start point of the segment.
+    xi1 : float
+        x-coordinate of end point of the segment.
+    yi0 : float
+        y-coordinate of start point of the segment.
+    yi1 : float
+        y-coordinate of end point of the segment.
+    bmin : float
+        Minimum relative distance from bpj1 at which slice should occur.
+    bmax1 : bool
+        Flag indicating whether the the relative distance along the segment bpj1-bpj should be limited to 1.
+
+    Returns
+    -------
+    a : np.ndarray
+        Array containing relative distance along each edge.
+    b : np.ndarray
+        Array containing relative distance along the segment for each edge.
+    slices : np.ndarray
+        Array containing a flag indicating whether the edge is sliced at a valid location.
+    """
+    dX = X1 - X0
+    dY = Y1 - Y0
+    dxi = xi1 - xi0
+    dyi = yi1 - yi0
+    det = dX * dyi - dY * dxi
+    det[det == 0] = 1e-10
+    a = (dyi * (xi0 - X0) - dxi * (yi0 - Y0)) / det  # along mesh edge
+    b = (dY * (xi0 - X0) - dX * (yi0 - Y0)) / det  # along bank line
+    if bmax1:
+        slices = np.nonzero((b > bmin) & (b <= 1) & (a >= 0) & (a <= 1))[0]
+    else:
+        slices = np.nonzero((b > bmin) & (a >= 0) & (a <= 1))[0]
+    a = a[slices]
+    b = b[slices]
+    return a, b, slices
+
+
+def enlarge(
+    old_array: np.ndarray,
+    new_shape: Tuple
+):
+    """
+    Copy the values of the old array to a new, larger array of specified shape.
+
+    Arguments
+    ---------
+    old_array : numpy.ndarray
+        Array containing the values.
+    new_shape : Tuple
+        New shape of the array.
+
+    Returns
+    -------
+    new_array : numpy.ndarray
+        Array of shape "new_shape" with the 'first entries filled by the same
+        values as contained in "old_array". The data type of the new array is
+        equal to that of the old array.
+    """
+    old_shape = old_array.shape
+    print("old: ", old_shape)
+    print("new: ", new_shape)
+    new_array = np.zeros(new_shape, dtype=old_array.dtype)
+    if len(new_shape)==1:
+        new_array[:old_shape[0]] = old_array
+    elif len(new_shape)==2:
+        new_array[:old_shape[0], :old_shape[1]] = old_array
+    return new_array
+
+
+def move_line(
+    xylines: np.ndarray, erosion_distance: np.ndarray, right_bank: bool
+) -> np.ndarray:
+    """
+    Shift a line of a variable distance sideways (positive shift away from centre line).
+
+    Chainage must be increasing along all lines. For a bank on the right side a
+    positive shift will move the line to the right. For a bank on the left side
+    a positive shift will move the line to the left.
+
+    Arguments
+    ---------
+    xylines : np.ndarray
+        Nx2 array containing the x- and y-coordinates of the line to be moved.
+    erosion_distance : np.ndarray
+        Distance over which to move the line sideways. A positive shift is
+        defined towards the right for the right bank, and towards the left for
+        the left bank.
+    right_bank : bool
+        Flag indicating whether line is on the right (or not).
+
+    Returns
+    -------
+    xylines_new : umpy.ndarray
+        Nx2 array containing the x- and y-coordinates of the moved line.
+    """
+    if right_bank:
+        xylines_new = _move_line_right(xylines, erosion_distance)
+    else:
+        xylines_rev = xylines[::-1, :]
+        dn_rev = erosion_distance[::-1]
+        xylines_new_rev = _move_line_right(xylines_rev, dn_rev)
+        xylines_new = xylines_new_rev[::-1, :]
+    return xylines_new
+
+
+def _move_line_right(xylines: np.ndarray, erosion_distance: np.ndarray) -> np.ndarray:
+    """
+    Shift a line of a variable distance sideways (positive shift to the right).
+
+    Arguments
+    ---------
+    xylines : np.ndarray
+        Nx2 array containing the x- and y-coordinates of the line to be moved.
+    dn0 : np.ndarray
+        Distance over which to move the line sideways. A positive shift is
+        defined towards the right when looking along the line.
+
+    Returns
+    -------
+    xylines_new : umpy.ndarray
+        Nx2 array containing the x- and y-coordinates of the moved line.
+    """
+    nsegments = len(erosion_distance)
+    colvec = (nsegments, 1)
+
+    # determine segment angle
+    dxy = xylines[1:, :] - xylines[:-1, :]
+    theta = np.arctan2(dxy[:, 1], dxy[:, 0])
+
+    # determine shift vector nxy for each segment
+    ds = np.sqrt((dxy ** 2).sum(axis=1))
+    nxy = dxy[:, ::-1] * [1, -1] * (erosion_distance / ds).reshape(colvec)
+
+    xylines_new = np.zeros((100, 2))
+    xylines_new[0] = xylines[0] + nxy[0]
+    ixy, xylines_new = _add_point(0, xylines_new, xylines[1] + nxy[0])
+    ixy, xylines_new = _add_point(ixy, xylines_new, xylines[1])
+
+    verbose = False
+    prec = 0.000001
+    ixy1: int
+    for iseg in range(1, nsegments):
+        dtheta = theta[iseg] - theta[iseg - 1]
+        if dtheta > math.pi:
+            dtheta = dtheta - 2 * math.pi
+        if verbose:
+            print("{}: current length of new bankline is {}".format(iseg, ixy))
+            print(
+                "{}: segment starting at {} to be shifted by {}".format(
+                    iseg, xylines[iseg], erosion_distance[iseg]
+                )
+            )
+            print("{}: change in direction quantified as {}".format(iseg, dtheta))
+
+        # create a polyline for the outline of the new segment
+        if erosion_distance[iseg] < prec:
+            # no erosion, so just a linear extension
+            if verbose:
+                print("{}: no shifting, just linear extension".format(iseg))
+            poly = np.row_stack([xylines[iseg + 1], xylines[iseg],])
+        elif dtheta <= 0:
+            # right bend
+            if -0.001 * math.pi < dtheta:
+                # almost straight
+                if verbose:
+                    print("{}: slight bend to right".format(iseg))
+                if erosion_distance[iseg] > erosion_distance[iseg]:
+                    poly = np.row_stack(
+                        [
+                            xylines[iseg + 1],
+                            xylines[iseg + 1] + nxy[iseg],
+                            xylines[iseg] + nxy[iseg],
+                            xylines[iseg] + nxy[iseg - 1],
+                            xylines[iseg - 1],
+                            ]
+                    )
+                else:
+                    poly = np.row_stack(
+                        [
+                            xylines[iseg + 1],
+                            xylines[iseg + 1] + nxy[iseg],
+                            xylines[iseg] + nxy[iseg],
+                            xylines[iseg - 1],
+                            ]
+                    )
+            else:
+                # more significant bend
+                if verbose:
+                    print("{}: bend to right".format(iseg))
+                poly = np.row_stack(
+                    [
+                        xylines[iseg + 1],
+                        xylines[iseg + 1] + nxy[iseg],
+                        xylines[iseg] + nxy[iseg],
+                        xylines[iseg],
+                        ]
+                )
+        elif erosion_distance[iseg - 1] < prec:
+            # left bend: previous segment isn't eroded, so nothing to connect to
+            if verbose:
+                print("{}: bend to left".format(iseg))
+            poly = np.row_stack(
+                [
+                    xylines[iseg + 1],
+                    xylines[iseg + 1] + nxy[iseg],
+                    xylines[iseg] + nxy[iseg],
+                    xylines[iseg],
+                    ]
+            )
+        else:
+            # left bend: connect it to the previous segment to avoid non eroded wedges
+            if verbose:
+                print("{}: bend to left".format(iseg))
+            poly = np.row_stack(
+                [
+                    xylines[iseg + 1],
+                    xylines[iseg + 1] + nxy[iseg],
+                    xylines[iseg] + nxy[iseg],
+                    xylines[iseg] + nxy[iseg - 1],
+                    xylines[iseg - 1],
+                    ]
+            )
+
+        nedges = poly.shape[0] - 1
+
+        # make a temporary copy of the last 20 nodes of the already shifted bankline
+        if ixy > 20:
+            X0 = xylines_new[(ixy - 20) : ixy, 0].copy()
+            Y0 = xylines_new[(ixy - 20) : ixy, 1].copy()
+            X1 = xylines_new[(ixy - 19) : (ixy + 1), 0].copy()
+            Y1 = xylines_new[(ixy - 19) : (ixy + 1), 1].copy()
+            ixy0 = ixy - 20
+        else:
+            X0 = xylines_new[:ixy, 0].copy()
+            Y0 = xylines_new[:ixy, 1].copy()
+            X1 = xylines_new[1 : ixy + 1, 0].copy()
+            Y1 = xylines_new[1 : ixy + 1, 1].copy()
+            ixy0 = 0
+
+        a = []
+        b = []
+        slices = []
+        n = []
+        # for each edge of the new polyline collect all intersections with the
+        # already shifted bankline ...
+        for i in range(nedges):
+            if (poly[i + 1] == poly[i]).all():
+                # polyline segment has no actual length, so skip it
+                pass
+            else:
+                # check for intersection
+                a2, b2, slices2 = _get_slices_ab(
+                    X0,
+                    Y0,
+                    X1,
+                    Y1,
+                    poly[i, 0],
+                    poly[i, 1],
+                    poly[i + 1, 0],
+                    poly[i + 1, 1],
+                    0,
+                    True,
+                )
+                # exclude the intersection if it's only at the very last point
+                # of the last segment
+                if i == nedges - 1:
+                    keep_mask = a2 < 1 - prec
+                    a2 = a2[keep_mask]
+                    b2 = b2[keep_mask]
+                    slices2 = slices2[keep_mask]
+                a.append(a2)
+                b.append(b2)
+                slices.append(slices2)
+                n.append(slices2 * 0 + i)
+
+        s = np.concatenate(slices)
+        if verbose:
+            print("{}: {} intersections detected".format(iseg, len(s)))
+        if len(s) == 0:
+            # no intersections found
+            if dtheta < 0:
+                # right bend (not straight)
+                if erosion_distance[iseg] > 0:
+                    cross = (xylines_new[ixy, 0] - xylines_new[ixy - 1, 0]) * nxy[
+                        iseg, 1
+                    ] - (xylines_new[ixy, 1] - xylines_new[ixy - 1, 1]) * nxy[iseg, 0]
+                else:
+                    cross = (xylines_new[ixy, 0] - xylines_new[ixy - 1, 0]) * dxy[
+                        iseg, 1
+                    ] - (xylines_new[ixy, 1] - xylines_new[ixy - 1, 1]) * dxy[iseg, 0]
+                if cross <= 0.0:
+                    # extended path turns right ... always add
+                    pass
+                else:
+                    # extended path turns left
+                    # we can probably ignore it, let's do so...
+                    # the only exception would be an eroded patch encompassing
+                    # all of the eroded bank line
+                    if verbose:
+                        print("{}: ignoring segment".format(iseg))
+                    continue
+            else:
+                # left bend or straight: always add ... just the rectangle of eroded material
+                pass
+            ixy1 = ixy
+            for n2 in range(min(nedges, 2), -1, -1):
+                if verbose:
+                    print("  adding point {}".format(poly[n2]))
+                ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
+            ixy = ixy1
+
+        else:
+            # one or more intersections found
+            a = np.concatenate(a)
+            b = np.concatenate(b)
+            n = np.concatenate(n)
+
+            # sort the intersections by distance along the already shifted bank line
+            d = s + a
+            sorted = np.argsort(d)
+            s = s[sorted] + ixy0
+            a = a[sorted]
+            b = b[sorted]
+            d = d[sorted]
+            n = n[sorted]
+
+            ixy1 = s[0]
+            if verbose:
+                print("{}: continuing new path at point {}".format(iseg, ixy1))
+            xytmp = xylines_new[ixy1 : ixy + 1].copy()
+            ixytmp = ixy1
+
+            inside = False
+            s_last = s[0]
+            n_last = nedges
+            for i in range(len(s)):
+                if verbose:
+                    print(
+                        "- intersection {}: new polyline edge {} crosses segment {} at {}".format(
+                            i, n[i], s[i], a[i]
+                        )
+                    )
+                if i == 0 or n[i] != nedges - 1:
+                    if inside:
+                        if verbose:
+                            print("  existing line is inside the new polygon")
+                        for n2 in range(n_last, n[i], -1):
+                            if verbose:
+                                print("  adding new point {}".format(poly[n2]))
+                            ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
+                    else:
+                        if verbose:
+                            print("  existing line is outside the new polygon")
+                        for s2 in range(s_last, s[i]):
+                            if verbose:
+                                print(
+                                    "  re-adding old point {}".format(
+                                        xytmp[s2 - ixytmp + 1]
+                                    )
+                                )
+                            ixy1, xylines_new = _add_point(
+                                ixy1, xylines_new, xytmp[s2 - ixytmp + 1]
+                            )
+                    pnt_intersect = poly[n[i]] + b[i] * (poly[n[i] + 1] - poly[n[i]])
+                    if verbose:
+                        print("  adding intersection point {}".format(pnt_intersect))
+                    ixy1, xylines_new = _add_point(ixy1, xylines_new, pnt_intersect, )
+                    n_last = n[i]
+                    s_last = s[i]
+                    if a[i] < prec:
+                        dPy = poly[n[i] + 1, 1] - poly[n[i], 1]
+                        dPx = poly[n[i] + 1, 0] - poly[n[i], 0]
+                        s2 = s[i] - ixy0
+                        dBy = Y1[s2] - Y0[s2]
+                        dBx = X1[s2] - X0[s2]
+                        inside = dPy * dBx - dPx * dBy > 0
+                    elif a[i] > 1 - prec:
+                        dPy = poly[n[i] + 1, 1] - poly[n[i], 1]
+                        dPx = poly[n[i] + 1, 0] - poly[n[i], 0]
+                        s2 = s[i] - ixy0 + 1
+                        if s2 > len(X0) - 1:
+                            inside = True
+                        else:
+                            dBy = Y1[s2] - Y0[s2]
+                            dBx = X1[s2] - X0[s2]
+                            inside = dPy * dBx - dPx * dBy > 0
+                    else:
+                        # line segment slices the edge somewhere in the middle
+                        inside = not inside
+                    if verbose:
+                        if inside:
+                            print("  existing line continues inside")
+                        else:
+                            print("  existing line continues outside")
+
+            if verbose:
+                print("- wrapping up after last intersection")
+            if inside:
+                if verbose:
+                    print("  existing line is inside the new polygon")
+                for n2 in range(n_last, -1, -1):
+                    if verbose:
+                        print("  adding new point {}".format(poly[n2]))
+                    ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
+            else:
+                if verbose:
+                    print("  existing line is inside the new polygon")
+                for s2 in range(s_last, len(xytmp) + ixytmp - 1):
+                    if verbose:
+                        print("  re-adding old point {}".format(xytmp[s2 - ixytmp + 1]))
+                    ixy1, xylines_new = _add_point(
+                        ixy1, xylines_new, xytmp[s2 - ixytmp + 1]
+                    )
+            ixy = ixy1
+        # if iseg == isegstop:
+        #     break
+    xylines_new = xylines_new[:ixy, :]
+
+    return xylines_new
+
+
+def _add_point(
+    ixy1: int, xy_in: np.ndarray, point: np.ndarray
+) -> Tuple[int, np.ndarray]:
+    """
+    Add the x,y-coordinates of a point to an array of x,y-coordinates if it differs from the last point.
+
+    Arguments
+    ---------
+    ixy1 : int
+        Index of last point in xy_in array
+    xy_in : np.ndarray
+        N x 2 array containing the x- and y-coordinates of points (partially filled)
+    point : np.ndarray
+        1 x 2 array containing the x- and y-coordinates of one point
+
+    Results
+    -------
+    ixy1 : int
+        Index of the new point in the xy_out array
+    xy_out : np.ndarray
+        Possibly extended copy of xy_in that includes the coordinates of point at ixy1
+    """
+    if (xy_in[ixy1] - point != 0).any():
+        ixy1 = ixy1 + 1
+        if ixy1 >= len(xy_in):
+            xy_out = enlarge(xy_in, (2 * ixy1, 2))
+        else:
+            xy_out = xy_in
+        xy_out[ixy1] = point
+    else:
+        xy_out = xy_in
+    return ixy1, xy_out
